+++ docs/conf.py
@@ -72,9 +72,9 @@ copyright = u'2015-2016 — Mozilla Services'
 # built documents.
 #
 # The short X.Y version.
-version = '4.2'
+version = '4.3'
 # The full version, including alpha/beta/rc tags.
-release = '4.2.0'
+release = '4.3.1'
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.