+++ kinto/__init__.py
@@ -13,7 +13,7 @@ from kinto.authorization import RouteFactory
 __version__ = pkg_resources.get_distribution(__package__).version
 
 # Implemented HTTP API Version
-HTTP_API_VERSION = '1.10'
+HTTP_API_VERSION = '1.12'
 
 # Main kinto logger
 logger = logging.getLogger(__name__)