+++ kinto/core/__init__.py
@@ -79,6 +79,7 @@ DEFAULT_SETTINGS = {
     'tm.annotate_user': False,  # Do annotate transactions with the user-id.
     'transaction_per_request': True,
     'userid_hmac_secret': '',
+    'version_json_path': 'version.json',
     'version_prefix_redirect_enabled': True,
     'trailing_slash_redirect_enabled': True,
     'multiauth.groupfinder': 'kinto.core.authorization.groupfinder',