+++ kinto/core/permission/__init__.py
@@ -133,8 +133,7 @@ class PermissionBase(object):
         return len(authorized & principals) > 0
 
     def get_object_permissions(self, object_id, permissions=None):
-        perms = self.get_objects_permissions([object_id], permissions)
-        return perms[0] if perms else {}
+        return self.get_objects_permissions([object_id], permissions)[0]
 
     def get_objects_permissions(self, objects_ids, permissions=None):
         """Return a list of mapping, for each object id specified, with the