+++ kinto/core/permission/memory.py
@@ -56,9 +56,12 @@ class Permission(PermissionBase):
 
     @synchronized
     def get_user_principals(self, user_id):
+        # Fetch the groups the user is in.
         user_key = 'user:%s' % user_id
         members = self._store.get(user_key, set())
-        return members
+        # Fetch the groups system.Authenticated is in.
+        group_authenticated = self._store.get('user:system.Authenticated', set())
+        return members | group_authenticated
 
     @synchronized
     def add_principal_to_ace(self, object_id, permission, principal):