+++ kinto/core/permission/postgresql/__init__.py
@@ -135,7 +135,8 @@ class Permission(PermissionBase):
         query = """
         SELECT principal
           FROM user_principals
-         WHERE user_id = :user_id;"""
+         WHERE user_id = :user_id
+            OR user_id = 'system.Authenticated';"""
         with self.client.connect(readonly=True) as conn:
             result = conn.execute(query, dict(user_id=user_id))
             results = result.fetchall()
@@ -300,12 +301,13 @@ class Permission(PermissionBase):
             rows = result.fetchall()
 
         groupby_id = OrderedDict()
+        for object_id in objects_ids:
+            groupby_id[object_id] = {}
         for row in rows:
             object_id, permission, principal = (row['object_id'],
                                                 row['permission'],
                                                 row['principal'])
-            permissions = groupby_id.setdefault(object_id, {})
-            permissions.setdefault(permission, set()).add(principal)
+            groupby_id[object_id].setdefault(permission, set()).add(principal)
         return list(groupby_id.values())
 
     def replace_object_permissions(self, object_id, permissions):