+++ kinto/core/permission/testing.py
@@ -140,6 +140,13 @@ class PermissionTest(object):
         retrieved = self.permission.get_user_principals(user_id2)
         self.assertEquals(retrieved, {principal2})
 
+    def test_authenticated_is_returned_for_everybody(self):
+        user_id = 'foo'
+        principal = 'bar'
+        self.permission.add_user_principal('system.Authenticated', principal)
+        retrieved = self.permission.get_user_principals(user_id)
+        self.assertEquals(retrieved, {principal})
+
     #
     # get_object_permission_principals()
     #
@@ -356,6 +363,14 @@ class PermissionTest(object):
             "write": {"user1", "user2"},
             "read": {"user3"}
         })
+
+    def test_objects_permissions_returns_empty_if_unknown(self):
+        self.permission.add_principal_to_ace('/url/a/id/1', 'write', 'user1')
+        self.permission.add_principal_to_ace('/url/a/id/3', 'read', 'user3')
+        objects_permissions = self.permission.get_objects_permissions([
+            '/url/a/id/1', '/abc', '/url/a/id/3'])
+        self.assertEqual(objects_permissions, [
+            {"write": {"user1"}}, {}, {"read": {"user3"}}])
 
     def test_object_permissions_return_empty_dict(self):
         self.assertDictEqual(self.permission.get_object_permissions('abc'), {})