+++ kinto/plugins/history/__init__.py
@@ -12,7 +12,13 @@ def includeme(config):
     # Activate end-points.
     config.scan('kinto.plugins.history.views')
 
+    # If StatsD is enabled, monitor execution time of listener.
+    listener = on_resource_changed
+    if config.registry.statsd:
+        key = 'plugins.history'
+        listener = config.registry.statsd.timer(key)(on_resource_changed)
+
     # Listen to every resources (except history)
-    config.add_subscriber(on_resource_changed, ResourceChanged,
+    config.add_subscriber(listener, ResourceChanged,
                           for_resources=('bucket', 'group',
                                          'collection', 'record'))