+++ kinto/plugins/history/views.py
@@ -2,6 +2,7 @@ import colander
 
 from kinto.core import resource
 from kinto.core.utils import instance_uri
+from kinto.core.storage import Filter
 
 
 class HistorySchema(resource.ResourceSchema):
@@ -31,3 +32,14 @@ class History(resource.ShareableResource):
     def get_parent_id(self, request):
         self.bucket_id = request.matchdict['bucket_id']
         return instance_uri(request, 'bucket', id=self.bucket_id)
+
+    def _extract_filters(self, queryparams=None):
+        filters = super(History, self)._extract_filters(queryparams)
+        filters_str_id = []
+        for filt in filters:
+            if filt.field in ('record_id', 'collection_id', 'bucket_id'):
+                if isinstance(filt.value, int):
+                    filt = Filter(filt.field, str(filt.value), filt.operator)
+            filters_str_id.append(filt)
+
+        return filters_str_id