+++ tests/test_authorization.py
@@ -1,7 +1,7 @@
 from kinto.core.testing import unittest
 
-from kinto.authorization import (get_object_type, build_permission_tuple,
-                                 build_permissions_set)
+from kinto.authorization import (_resource_endpoint, _relative_object_uri,
+                                 _inherited_permissions)
 
 
 class PermissionInheritanceTest(unittest.TestCase):
@@ -11,92 +11,85 @@ class PermissionInheritanceTest(unittest.TestCase):
     bucket_uri = '/buckets/blog'
     invalid_uri = 'invalid object id'
 
-    def test_get_object_type_return_right_type_for_key(self):
-        self.assertEqual(get_object_type(self.record_uri), 'record')
-        self.assertEqual(get_object_type(self.collection_uri), 'collection')
-        self.assertEqual(get_object_type(self.bucket_uri), 'bucket')
-        self.assertEqual(get_object_type(self.group_uri), 'group')
-        self.assertIsNone(get_object_type(self.invalid_uri))
+    def test_resource_endpoint_return_right_type_for_key(self):
+        self.assertEqual(_resource_endpoint(self.record_uri), ('record', False))
+        self.assertEqual(_resource_endpoint(self.collection_uri), ('collection', False))
+        self.assertEqual(_resource_endpoint(self.bucket_uri), ('bucket', False))
+        self.assertEqual(_resource_endpoint(self.group_uri), ('group', False))
+        self.assertRaises(ValueError, _resource_endpoint, self.invalid_uri)
 
-    def test_get_object_type_return_right_type_for_children_collection(self):
-        object_type = get_object_type(self.collection_uri + '/records')
-        self.assertEqual(object_type, 'collection')
-        object_type = get_object_type(self.bucket_uri + '/collections')
-        self.assertEqual(object_type, 'bucket')
-        object_type = get_object_type(self.bucket_uri + '/groups')
-        self.assertEqual(object_type, 'bucket')
+    def test_resource_endpoint_return_right_type_for_children_collection(self):
+        self.assertEqual(_resource_endpoint(self.collection_uri + '/records'),
+                         ('collection', True))
+        self.assertEqual(_resource_endpoint(self.bucket_uri + '/collections'),
+                         ('bucket', True))
+        self.assertEqual(_resource_endpoint(self.bucket_uri + '/groups'),
+                         ('bucket', True))
 
-    def test_build_perm_set_uri_can_construct_parents_set_uris(self):
-        obj_parts = self.record_uri.split('/')
+    def test_related_object_uri_can_construct_parents_set_uris(self):
         # Can build record_uri from obj_parts
-        self.assertEqual(
-            build_permission_tuple('record', 'write', obj_parts),
-            (self.record_uri, 'write'))
+        self.assertEqual(_relative_object_uri('record', self.record_uri),
+                         self.record_uri)
 
         # Can build collection_uri from obj_parts
-        self.assertEqual(
-            build_permission_tuple('collection', 'records:create', obj_parts),
-            (self.collection_uri, 'records:create'))
+        self.assertEqual(_relative_object_uri('collection', self.record_uri),
+                         self.collection_uri)
 
         # Can build bucket_uri from obj_parts
-        self.assertEqual(build_permission_tuple(
-            'bucket', 'groups:create', obj_parts),
-            (self.bucket_uri, 'groups:create'))
+        self.assertEqual(_relative_object_uri('bucket', self.record_uri),
+                         self.bucket_uri)
 
         # Can build group_uri from group obj_parts
-        obj_parts = self.group_uri.split('/')
-        self.assertEqual(build_permission_tuple(
-            'group', 'read', obj_parts),
-            (self.group_uri, 'read'))
+        self.assertEqual(_relative_object_uri('group', self.group_uri),
+                         self.group_uri)
 
         # Can build bucket_uri from group obj_parts
-        obj_parts = self.group_uri.split('/')
-        self.assertEqual(build_permission_tuple(
-            'bucket', 'write', obj_parts),
-            (self.bucket_uri, 'write'))
+        self.assertEqual(_relative_object_uri('bucket', self.group_uri),
+                         self.bucket_uri)
 
-    def test_build_perm_set_supports_buckets_named_collections(self):
-        uri = '/buckets/collections'
-        self.assertEquals(build_permissions_set(uri, 'write'),
-                          set([(uri, 'write')]))
-
-    def test_build_permission_tuple_fail_construct_children_set_uris(self):
-        obj_parts = self.bucket_uri.split('/')
+    def test_relative_object_uri_fail_construct_children_set_uris(self):
         # Cannot build record_uri from bucket obj_parts
         self.assertRaises(ValueError,
-                          build_permission_tuple,
-                          'record', 'write', obj_parts)
+                          _relative_object_uri,
+                          'record', self.bucket_uri)
 
         # Cannot build collection_uri from obj_parts
         self.assertRaises(ValueError,
-                          build_permission_tuple,
-                          'collection', 'write', obj_parts)
+                          _relative_object_uri,
+                          'collection', self.bucket_uri)
 
         # Cannot build bucket_uri from empty obj_parts
         self.assertRaises(ValueError,
-                          build_permission_tuple,
-                          'collection', 'write', [])
+                          _relative_object_uri,
+                          'collection', '')
 
-    def test_build_permission_tuple_fail_on_wrong_type(self):
-        obj_parts = self.record_uri.split('/')
-        self.assertRaises(ValueError,
-                          build_permission_tuple,
-                          'schema', 'write', obj_parts)
+    def test_relative_object_uri_fail_on_wrong_type(self):
+        self.assertRaises(KeyError,
+                          _relative_object_uri,
+                          'schema', self.record_uri)
 
-    def test_get_perm_keys_for_bucket_permission(self):
+    def test_inherited_permissions_supports_buckets_named_collections(self):
+        uri = '/buckets/collections'
+        self.assertEquals(set(_inherited_permissions(uri, 'write')),
+                          set([(uri, 'write')]))
+
+    def test_inherited_permissions_for_bucket_permission(self):
         # write
         self.assertEquals(
-            build_permissions_set(self.bucket_uri, 'write'),
-            set([(self.bucket_uri, 'write')]))
+            _inherited_permissions(self.bucket_uri, 'write'),
+            [(self.bucket_uri, 'write')])
         # read
         self.assertEquals(
-            build_permissions_set(self.bucket_uri, 'read'),
-            set([(self.bucket_uri, 'write'), (self.bucket_uri, 'read')]))
+            set(_inherited_permissions(self.bucket_uri, 'read')),
+            set([(self.bucket_uri, 'write'),
+                 (self.bucket_uri, 'read'),
+                 (self.bucket_uri, 'collection:create'),
+                 (self.bucket_uri, 'group:create')]))
 
         # group:create
         groups_uri = self.bucket_uri + '/groups'
         self.assertEquals(
-            build_permissions_set(groups_uri, 'group:create'),
+            set(_inherited_permissions(groups_uri, 'group:create')),
             set(
                 [(self.bucket_uri, 'write'),
                  (self.bucket_uri, 'group:create')])
@@ -105,55 +98,56 @@ class PermissionInheritanceTest(unittest.TestCase):
         # collection:create
         collections_uri = self.bucket_uri + '/collections'
         self.assertEquals(
-            build_permissions_set(collections_uri, 'collection:create'),
+            set(_inherited_permissions(collections_uri, 'collection:create')),
             set([(self.bucket_uri, 'write'),
                  (self.bucket_uri, 'collection:create')]))
 
-    def test_build_permissions_set_for_group_permission(self):
+    def test_inherited_permissions_for_group_permission(self):
         # write
         self.assertEquals(
-            build_permissions_set(self.group_uri, 'write'),
-            set([(self.bucket_uri, 'write'),
-                 (self.group_uri, 'write')]))
+            _inherited_permissions(self.group_uri, 'write'),
+            [(self.group_uri, 'write'),
+             (self.bucket_uri, 'write')])
         # read
         self.assertEquals(
-            build_permissions_set(self.group_uri, 'read'),
+            set(_inherited_permissions(self.group_uri, 'read')),
             set([(self.bucket_uri, 'write'),
                  (self.bucket_uri, 'read'),
                  (self.group_uri, 'write'),
                  (self.group_uri, 'read')]))
 
-    def test_build_permissions_set_for_collection_permission(self):
+    def test_inherited_permissions_for_collection_permission(self):
         # write
         self.assertEquals(
-            build_permissions_set(self.collection_uri, 'write'),
-            set([(self.bucket_uri, 'write'),
-                 (self.collection_uri, 'write')]))
+            _inherited_permissions(self.collection_uri, 'write'),
+            [(self.collection_uri, 'write'),
+             (self.bucket_uri, 'write')])
         # read
         self.assertEquals(
-            build_permissions_set(self.collection_uri, 'read'),
+            set(_inherited_permissions(self.collection_uri, 'read')),
             set([(self.bucket_uri, 'write'),
                  (self.bucket_uri, 'read'),
                  (self.collection_uri, 'write'),
-                 (self.collection_uri, 'read')]))
+                 (self.collection_uri, 'read'),
+                 (self.collection_uri, 'record:create')]))
         # records:create
         records_uri = self.collection_uri + '/records'
         self.assertEquals(
-            build_permissions_set(records_uri, 'record:create'),
+            set(_inherited_permissions(records_uri, 'record:create')),
             set([(self.bucket_uri, 'write'),
                  (self.collection_uri, 'write'),
                  (self.collection_uri, 'record:create')]))
 
-    def test_build_permissions_set_for_record_permission(self):
+    def test_inherited_permissions_for_record_permission(self):
         # write
         self.assertEquals(
-            build_permissions_set(self.record_uri, 'write'),
+            set(_inherited_permissions(self.record_uri, 'write')),
             set([(self.bucket_uri, 'write'),
                  (self.collection_uri, 'write'),
                  (self.record_uri, 'write')]))
         # read
         self.assertEquals(
-            build_permissions_set(self.record_uri, 'read'),
+            set(_inherited_permissions(self.record_uri, 'read')),
             set([(self.bucket_uri, 'write'),
                  (self.bucket_uri, 'read'),
                  (self.collection_uri, 'write'),
@@ -161,6 +155,6 @@ class PermissionInheritanceTest(unittest.TestCase):
                  (self.record_uri, 'write'),
                  (self.record_uri, 'read')]))
 
-    def test_build_permissions_set_returns_empty_set_if_doesnt_know(self):
-        permissions = build_permissions_set('/buckets', 'read')
-        self.assertEquals(permissions, set())
+    def test_inherited_permissions_for_list_of_buckets(self):
+        permissions = _inherited_permissions('/buckets', 'read')
+        self.assertEquals(permissions, [])